apiserver/requirements.txt
# Ginkgo API Server Requirements

# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# 数据验证
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 认证
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6

# Kafka集成
kafka-python>=2.0.2

# WebSocket
websockets>=12.0

# 数据库（通过Ginkgo核心间接使用）
# sqlalchemy、pymongo、clickhouse-driver、redis等由Ginkgo核心提供

# 日志
python-json-logger>=2.0.7

# 工具
python-dotenv>=1.0.0
